refactor(error-notification): storageをドメイン知識ごとに整理
- models/storage_location.py → storage_object.py に移動
- interfaces/object_repository_interface.py → storage_interface.py に移動
- __init__.py のエクスポートを更新
