+++ lambda/error-notification/src/domain/interfaces/__init__.py
@@ -1,9 +1,7 @@
 """ドメインインターフェース"""
 
 from .notifier_interface import INotifier
-from .object_repository_interface import IObjectRepository
 
 __all__ = [
-    "IObjectRepository",
     "INotifier",
 ]
