+++ lambda/error-notification/src/domain/models/__init__.py
@@ -1,9 +1,7 @@
 """ドメインモデル"""
 
 from .notification_message import NotificationMessage
-from .storage_location import StorageLocation
 
 __all__ = [
-    "StorageLocation",
     "NotificationMessage",
 ]
